+++ src/OrleansConsulUtils/ConsulUtilsHostingExtensions.cs
@@ -1,19 +1,37 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
-using Orleans;
 using Orleans.ConsulUtils.Configuration;
-using Orleans.Runtime.Configuration;
-using Orleans.Runtime.Host;
 using Orleans.Runtime.Membership;
 
-namespace Orleans.Runtime.Hosting 
+namespace Orleans.Hosting
 {
-    public static class ConsulUtilsServiceCollectionExtensions
+    public static class ConsulUtilsHostingExtensions 
     {
+        /// <summary>
+        /// Configure siloHostBuilder to use ConsulMembership
+        /// </summary>
+        /// <param name="builder"></param>
+        /// <param name="configureOptions"></param>
+        /// <returns></returns>
+        public static ISiloHostBuilder UseConsulMembership(this ISiloHostBuilder builder,
+            Action<ConsulMembershipOptions> configureOptions)
+        {
+            return builder.ConfigureServices(services => services.UseConsulMembership(configureOptions));
+        }
+
+        /// <summary>
+        /// Configure siloHostBuilder to use ConsulMembership
+        /// </summary>
+        /// <param name="builder"></param>
+        /// <param name="configuration"></param>
+        /// <returns></returns>
+        public static ISiloHostBuilder UseConsulMembership(this ISiloHostBuilder builder,
+            IConfiguration configuration)
+        {
+            return builder.ConfigureServices(services => services.UseConsulMembership(configuration));
+        }
+
         /// <summary>
         /// Configure DI container with ConsuleBasedMemebership
         /// </summary>
