+++ test/AWSUtils.Tests/LivenessTests.cs
@@ -10,7 +10,6 @@ using Amazon.Runtime;
 using Microsoft.Extensions.Logging.Abstractions;
 using Orleans.Hosting;
 using Orleans.Runtime.Configuration;
-using Orleans.Runtime.Hosting;
 using Orleans.TestingHost;
 using Orleans.TestingHost.Utils;
 using UnitTests.MembershipTests;
